Move signStates function out of test-helpers
